Implements rough version of GraphQL endpoint for the UserSchema
Redefine the user routes with a "/graphql" endpoint, serving a
dummy version of the schema which does not use the actor system
for resolution. This is based on the example in the sangria docs.

Also, get requests to this endpoint return a static HTML file with GraphiQL,
allowing creation and execution of queries directly from the browser.