Adds create user mutation
Define the resolver as a class in the same file as the schema.
This class keeps things simple, while exposing the implicit
execution context so that the resolve function defined in
the mutation field can use map, flatMap, etc on futures.

Check that the id of the new user is unique as a seperate ask
within the resolve function of the mutation, instead of changing
the commands of the user registry. This keeps the registry API
uncomplicated.